Fix for better italian writing
Italian writing tipically use grave accent for all accented words like
"però", "più", "lì" etc. An exception exists for the "e" letter that can
be acute like in "perché" or "trentatré" or grave like in "purè" or "è".
I fixed all the vowel mappings and added the possibility to type an "é"
using the R key (which being next to the E key should feel natural).

Overall I made this changes:
AltGr + a: à    (instead of á)
AltGr + e: è    (instead of é)
AltGr + i: ì    (instead of í)
AltGr + o: ò    (instead of ó)
AltGr + u: ù    (instead of ú)
AltGr + r: é    (instead of ®)
AltGr + R: ®    (it was unassigned)